+++ tests_basic/test_planar_imaging.py
@@ -264,7 +264,7 @@ class GeneralTests(TestCase):
 class PlanarPhantomMixin(QuaacTestBase, CloudFileMixin, PlotlyTestMixin):
     klass: Callable
     dir_path = ["planar_imaging"]
-    mtf_50 = None
+    mtf_values = {}
     invert = False
     ssd = "auto"
     median_contrast = None
@@ -309,15 +309,16 @@ class PlanarPhantomMixin(QuaacTestBase, CloudFileMixin, PlotlyTestMixin):
         super().tearDownClass()
 
     def test_bad_inversion_recovers(self):
-        if self.mtf_50 is None:
-            self.skipTest("mtf_50 not available")
+        if not self.mtf_values:
+            self.skipTest("mtf_values not available")
         instance = self.create_instance()
         instance.image.invert()
         instance.analyze(ssd=self.ssd, invert=self.invert)
         # check that the MTF is the expected value. This is a surrogate for the angle being wrong
-        self.assertAlmostEqual(
-            self.mtf_50, instance.mtf.relative_resolution(50), delta=0.3
-        )
+        for percent, expected in self.mtf_values.items():
+            self.assertAlmostEqual(
+                expected, instance.mtf.relative_resolution(percent), delta=0.3
+            )
 
     def test_plotting(self):
         self.instance.plot_analyzed_image()
@@ -334,11 +335,12 @@ class PlanarPhantomMixin(QuaacTestBase, CloudFileMixin, PlotlyTestMixin):
         save_file(self.instance.publish_pdf)
 
     def test_mtf(self):
-        if self.mtf_50 is None:
-            self.skipTest("mtf_50 not available")
-        self.assertAlmostEqual(
-            self.mtf_50, self.instance.mtf.relative_resolution(50), delta=0.3
-        )
+        if not self.mtf_values:
+            self.skipTest("mtf_values not available")
+        for percent, expected in self.mtf_values.items():
+            self.assertAlmostEqual(
+                expected, self.instance.mtf.relative_resolution(percent), delta=0.3
+            )
 
     def test_rois_seen(self):
         if self.rois_seen is None:
@@ -390,7 +392,7 @@ class LeedsMixin(PlanarPhantomMixin):
 
 
 class LeedsDemo(LeedsMixin, TestCase):
-    mtf_50 = 1.5
+    mtf_values = {50: 1.5}
     piu = 91.8
 
     def test_demo(self):
@@ -398,27 +400,27 @@ class LeedsDemo(LeedsMixin, TestCase):
 
 
 class LeedsCCW(LeedsMixin, TestCase):
-    mtf_50 = 1.5
+    mtf_values = {50: 1.5}
     file_name = "Leeds_ccw.dcm"
     piu = 93.5
 
 
 class Leeds45Deg(LeedsMixin, TestCase):
-    mtf_50 = 1.9
+    mtf_values = {50: 1.9}
     ssd = "auto"
     file_name = "Leeds-45deg.dcm"
     piu = 95.5
 
 
 class LeedsDirtyEdges(LeedsMixin, TestCase):
-    mtf_50 = 1.53
+    mtf_values = {50: 1.53}
     ssd = "auto"
     file_name = "Leeds-dirty-edges.dcm"
     piu = 96.8
 
 
 class LeedsOffsetHighRes(LeedsMixin, TestCase):
-    mtf_50 = 1.85
+    mtf_values = {50: 1.85}
     ssd = "auto"
     file_name = "Leeds_offset_high_res_rois.dcm"
     piu = 89.3
@@ -426,7 +428,7 @@ class LeedsOffsetHighRes(LeedsMixin, TestCase):
 
 class LeedsBlue(LeedsMixin, TestCase):
     klass = LeedsTORBlue
-    mtf_50 = 1.5
+    mtf_values = {50: 1.5}
     ssd = "auto"
     file_name = "Leeds_Blue.dcm"
     piu = 97
@@ -435,7 +437,7 @@ class LeedsBlue(LeedsMixin, TestCase):
 
 class LeedsBlueRotated(LeedsMixin, TestCase):
     klass = LeedsTORBlue
-    mtf_50 = 1.5
+    mtf_values = {50: 1.5}
     ssd = "auto"
     file_name = "Leeds_Blue.dcm"
     piu = 97
@@ -450,7 +452,7 @@ class LeedsBlueRotated(LeedsMixin, TestCase):
 
 @skip("Phantom appears distorted. MTF locations are different than other phantoms")
 class LeedsClosedBlades(LeedsMixin, TestCase):
-    mtf_50 = 1.3
+    mtf_values = {50: 1.3}
     ssd = "auto"
     file_name = "Leeds-closed-blades.dcm"
 
@@ -458,7 +460,7 @@ class LeedsClosedBlades(LeedsMixin, TestCase):
 class LeedsACB1(LeedsMixin, TestCase):
     dir_path = ["planar_imaging", "Leeds", "ACB 1"]
     file_name = "1.dcm"
-    mtf_50 = 1.42
+    mtf_values = {50: 1.42}
     piu = 98.80
 
 
@@ -466,13 +468,13 @@ class LeedsBadInversion(LeedsMixin, TestCase):
     """Radmachine image where inversion was bad. pylinac should be able to correct"""
 
     file_name = "Leeds bad inversion.dcm"
-    mtf_50 = 1.21
+    mtf_values = {50: 1.21}
     piu = 97.86
 
 
 class SIQC3Demo(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQC3
-    mtf_50 = 0.53
+    mtf_values = {50: 0.53}
     rois_seen = 5
     piu = 98
 
@@ -483,7 +485,7 @@ class SIQC3Demo(PlanarPhantomMixin, TestCase):
 class SIQC3_1(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQC3
     file_name = "QC3-2.5MV.dcm"
-    mtf_50 = 1.19
+    mtf_values = {70: 0.571}
     rois_seen = 5
     piu = 91.8
 
@@ -491,7 +493,7 @@ class SIQC3_1(PlanarPhantomMixin, TestCase):
 class SIQC3_2(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQC3
     file_name = "QC3-2.5MV-2.dcm"
-    mtf_50 = 1.16
+    mtf_values = {70: 0.576}
     ssd = 1000
     rois_seen = 5
     piu = 91.8
@@ -707,7 +709,7 @@ class Elekta10MU(ElektaLasVegasMixin, TestCase):
 
 class DoselabMVDemo(PlanarPhantomMixin, TestCase):
     klass = DoselabMC2MV
-    mtf_50 = 0.54
+    mtf_values = {50: 0.54}
     piu = 48.7
 
     def test_demo(self):
@@ -727,7 +729,7 @@ class DoseLabMVRotated(PlanarPhantomMixin, TestCase):
 
 class DoselabkVDemo(PlanarPhantomMixin, TestCase):
     klass = DoselabMC2kV
-    mtf_50 = 2.0
+    mtf_values = {50: 2.0}
     piu = 4.2
 
     def test_demo(self):
@@ -751,7 +753,7 @@ class DoselabkV70kVp(PlanarPhantomMixin, TestCase):
     klass = DoselabMC2kV
     dir_path = ["planar_imaging", "Doselab MC2"]
     file_name = "DL kV 70kVp.dcm"
-    mtf_50 = 1.14
+    mtf_values = {50: 1.14}
     piu = 0
 
     def test_window_ceiling(self):
@@ -774,7 +776,7 @@ class DoseLabkVRotated(PlanarPhantomMixin, TestCase):
 
 class SNCkVDemo(PlanarPhantomMixin, TestCase):
     klass = SNCkV
-    mtf_50 = 1.76
+    mtf_values = {50: 1.76}
     median_contrast = 0.17
     median_cnr = 69.4
     piu = 98.7
@@ -787,7 +789,7 @@ class SNCMVDemo(PlanarPhantomMixin, TestCase):
     klass = SNCMV
     median_cnr = 81
     median_contrast = 0.21
-    mtf_50 = 0.43
+    mtf_values = {50: 0.43}
     piu = 98.4
 
     def test_demo(self):
@@ -796,7 +798,7 @@ class SNCMVDemo(PlanarPhantomMixin, TestCase):
 
 class SNCMV12510_6MV1(PlanarPhantomMixin, TestCase):
     klass = SNCMV12510
-    mtf_50 = 0.91
+    mtf_values = {50: 0.91}
     median_contrast = 0.254
     median_cnr = 65.34
     dir_path = ["planar_imaging", "SNC MV Old"]
@@ -809,7 +811,7 @@ class SNCMV12510_6MV1(PlanarPhantomMixin, TestCase):
 
 class SNCMV12510_6MV2(PlanarPhantomMixin, TestCase):
     klass = SNCMV12510
-    mtf_50 = 0.85
+    mtf_values = {50: 0.85}
     median_contrast = 0.255
     median_cnr = 66.43
     dir_path = ["planar_imaging", "SNC MV Old"]
@@ -821,7 +823,7 @@ class SNCMV12510_Jig(PlanarPhantomMixin, TestCase):
     """Phantom where the jig is touching and gets in the way of analysis"""
 
     klass = SNCMV12510
-    mtf_50 = 0.92
+    mtf_values = {50: 0.92}
     median_contrast = 0.23
     median_cnr = 58.6
     dir_path = ["planar_imaging", "SNC MV Old"]
@@ -833,7 +835,7 @@ class SNCMV12510_HighRes(PlanarPhantomMixin, TestCase):
     """The HighRes and LowRes tests use the same phantom but different imaging resolution"""
 
     klass = SNCMV12510
-    mtf_50 = 0.92
+    mtf_values = {50: 0.92}
     median_contrast = 0.23
     median_cnr = 63.5
     dir_path = ["planar_imaging", "SNC MV Old"]
@@ -845,7 +847,7 @@ class SNCMV12510_LowRes(PlanarPhantomMixin, TestCase):
     """The HighRes and LowRes tests use the same phantom but different imaging resolution"""
 
     klass = SNCMV12510
-    mtf_50 = 0.71
+    mtf_values = {50: 0.71}
     median_contrast = 0.23
     median_cnr = 61.8
     dir_path = ["planar_imaging", "SNC MV Old"]
@@ -857,7 +859,7 @@ class IBAPrimusDemo(PlanarPhantomMixin, TestCase):
     klass = IBAPrimusA
     dir_path = ["planar_imaging", "PrimusL"]
     file_name = "Demo.dcm"
-    mtf_50 = 1.66
+    mtf_values = {50: 1.33}
     ssd = 1395
     median_cnr = 1084.4
     median_contrast = 0.62
@@ -910,7 +912,7 @@ class IBAPrimusFarSSD(PlanarPhantomMixin, TestCase):
     klass = IBAPrimusA
     dir_path = ["planar_imaging", "PrimusL"]
     file_name = "Primus_farSSD.dcm"
-    mtf_50 = 2.33
+    mtf_values = {50: 1.95}
     ssd = 2790
     median_cnr = 3990
     median_contrast = 0.6
@@ -919,7 +921,7 @@ class IBAPrimusFarSSD(PlanarPhantomMixin, TestCase):
 
 class SIQCkVDemo(PlanarPhantomMixin, TestCase):
     klass = StandardImagingQCkV
-    mtf_50 = 1.81
+    mtf_values = {50: 1.81}
     rois_seen = 5
     piu = 96.7
 
@@ -929,7 +931,7 @@ class SIQCkVDemo(PlanarPhantomMixin, TestCase):
 
 class PTWEPIDDemo(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.79
+    mtf_values = {50: 0.79}
     piu = 95.4
 
     def test_demo(self):
@@ -938,7 +940,7 @@ class PTWEPIDDemo(PlanarPhantomMixin, TestCase):
 
 class PTWEPIDQC1(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.79
+    mtf_values = {50: 0.79}
     rois_seen = 9
     median_contrast = 0.26
     median_cnr = 40.9
@@ -949,7 +951,7 @@ class PTWEPIDQC1(PlanarPhantomMixin, TestCase):
 
 class PTWEPID15MV(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.5
+    mtf_values = {50: 0.5}
     rois_seen = 9
     median_contrast = 0.17
     median_cnr = 26.7
@@ -960,7 +962,7 @@ class PTWEPID15MV(PlanarPhantomMixin, TestCase):
 
 class PTWEPID6xHigh(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.79
+    mtf_values = {50: 0.79}
     rois_seen = 9
     median_contrast = 0.28
     median_cnr = 72.1
@@ -971,7 +973,7 @@ class PTWEPID6xHigh(PlanarPhantomMixin, TestCase):
 
 class PTWEPID6xHighQuality(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.79
+    mtf_values = {50: 0.79}
     rois_seen = 9
     median_contrast = 0.254
     median_cnr = 37.9
@@ -982,7 +984,7 @@ class PTWEPID6xHighQuality(PlanarPhantomMixin, TestCase):
 
 class PTWEPIDTB3(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.79
+    mtf_values = {50: 0.79}
     rois_seen = 9
     median_contrast = 0.31
     median_cnr = 43.2
@@ -993,7 +995,7 @@ class PTWEPIDTB3(PlanarPhantomMixin, TestCase):
 
 class PTWEPIDTB4(PlanarPhantomMixin, TestCase):
     klass = PTWEPIDQC
-    mtf_50 = 0.79
+    mtf_values = {50: 0.79}
     rois_seen = 9
     median_contrast = 0.30
     median_cnr = 39.1